requirements.txt
django~=1.11
djangorestframework~=3.3
django-rest-auth~=0.7
python-social-auth[django]~=0.3
python3-openid==3.0.10
IGitt==0.4.1.dev20171013112039
drfdocs~=0.0.11
django-cors-headers~=1.3.1
gunicorn~=19.7.1
gevent~=1.2.2
psycopg2~=2.7.1
bugspots3~=0.3.0
attrs~=17.2.0
django-pglocks~=1.0.2
raven~=6.2.1
django-rq~=0.9.6
rq-scheduler~=0.7.0
